requirements.txt
torch>=2.1
numpy>=1.23
pyyaml>=6.0
tqdm>=4.66
